Refactor sender workers, reduce number of workers
Removed the "sender" workers, that would only have a single method to
send an email. These methods were moved to the workers that called them.
Less workers, more organized code.
