+++ spec/workers/participant_confirmations_worker_spec.rb
@@ -8,9 +8,8 @@ require 'spec_helper'
 
 describe ParticipantConfirmationsWorker, type: :worker do
   let(:worker) { ParticipantConfirmationsWorker }
-  let(:sender) { ParticipantConfirmationsSenderWorker }
   let(:queue) { Queue::High }
-  let(:params) {{"method"=>:perform, "class"=>sender.to_s}}
+  let(:params) {{"method"=>:confirmation_sender, "class"=>ParticipantConfirmationsWorker.to_s}}
 
   describe "#perform" do
     before { ParticipantConfirmation.delete_all }
@@ -28,7 +27,24 @@ describe ParticipantConfirmationsWorker, type: :worker do
       it { expect(queue).not_to have_queued(params, pc2.id) }
       it { expect(queue).not_to have_queued(params, pc4.id) }
     end
-
   end
 
+  describe "#confirmation_sender" do
+    let(:pc1) { FactoryGirl.create(:participant, email: 'abc@def.cam', owner: nil).participant_confirmation }
+    let(:pc2) { FactoryGirl.create(:participant, email: 'def@abc.com', owner: nil).participant_confirmation }
+    before(:each) {
+      pc2.update_attribute(:email_sent_at, Time.now)
+      worker.confirmation_sender(pc1.id)
+      worker.confirmation_sender(pc2.id)
+    }
+
+    it { ParticipantConfirmationMailer.should have_queue_size_of(1) }
+    it { ParticipantConfirmationMailer.should have_queued(:confirmation_email, pc1.id).in(:mailer) }
+    it { ParticipantConfirmationMailer.should_not have_queued(:confirmation_email, pc2.id).in(:mailer) }
+    it { pc1.reload.email_sent_at.should_not be_nil }
+    it { pc2.reload.email_sent_at.should_not be_nil }
+  end
+
+  # might happen if an admin removes a participant before the notification is sent
+  it "doesn't break if a ParticipantConfirmation has no participant associated"
 end
